# GS: Improve autoflush rect checks