### Description of Changes
Improves the rect calculation of the current draw and compares it to the last texture position from the previous draw to check for page crossings. 

### Rationale behind Changes
Reduces change of false positives and thus lower autoflush performance. 

### Suggested Testing Steps
Test games that require Autoflush, make sure they still work good.


Also unbreaks Sonic Unleashed when Autoflush is enabled in hardware mode.

Performance metrics compared to Master using GS Dumps at Native x3 with Autoflush enabled unless otherwise stated:

Black goes from 224fps to 400fps
GoW 2 goes from 150fps to 300fps
Jak 2 goes from 60fps to 71fps. At Native this translates to 144fps to 160fps
Scarface goes from 200fps to 468fps
Sniper Elite goes from 107fps to 287fps