@AhayriSG 2022-04-02
SW - Auto Flush: On
![gs_20220402135003_UEFA Champions League 2006-2007 _NTSC-U__SLUS-21581](https://user-images.githubusercontent.com/59465544/161380059-11b0026d-a46f-4929-98f8-714a3859689a.png)
SW - Auto Flush: Off
![gs_20220402135019_UEFA Champions League 2006-2007 _NTSC-U__SLUS-21581](https://user-images.githubusercontent.com/59465544/161380063-5c4b6c2a-f74c-4f25-8f22-adf77930205e.png)
Master (SW - Auto Flush: On)
https://wiki.pcsx2.net/File:Gs_20220331205217_UEFA_Champions_League_2006-2007_NTSC-U_SLUS-21581.png