@refractionpcsx2 2022-04-02
Thanks :) I think I messed up something with the vertex kick functions lol