@refractionpcsx2 2022-04-02
Sorry about that, I reverted the experimental change I did last night which was causing problems, once I fixed it (but kept it in place) the was no advantage to it anymore, so I just got rid.